IGNITE-28819 Move IgniteUtils.cast/unwrap to CommonUtils
Relocate the exception helpers cast(Throwable) and unwrap(Throwable) from
ignite-core's IgniteUtils to ignite-commons's CommonUtils so they are available
to modules that must not depend on ignite-core (GridFutureAdapter uses
U.cast). Both depend only on commons types (GridClosureException,
IgniteCheckedException). Since IgniteUtils extends CommonUtils, existing
U.cast / U.unwrap call sites keep working via inheritance, with no caller
changes.
