+++ modules/core/src/main/java/org/apache/ignite/internal/util/IgniteUtils.java
@@ -192,7 +192,6 @@ import org.apache.ignite.internal.transactions.IgniteTxRollbackCheckedException;
 import org.apache.ignite.internal.transactions.IgniteTxTimeoutCheckedException;
 import org.apache.ignite.internal.util.future.GridFutureAdapter;
 import org.apache.ignite.internal.util.future.IgniteFutureImpl;
-import org.apache.ignite.internal.util.lang.GridClosureException;
 import org.apache.ignite.internal.util.lang.GridPeerDeployAware;
 import org.apache.ignite.internal.util.lang.IgniteThrowableFunction;
 import org.apache.ignite.internal.util.typedef.C1;
@@ -4866,49 +4865,6 @@ public abstract class IgniteUtils extends CommonUtils {
      */
     public static JMException jmException(Throwable e) {
         return new JMException(e.getMessage());
-    }
-
-    /**
-     * Unwraps closure exceptions.
-     *
-     * @param t Exception.
-     * @return Unwrapped exception.
-     */
-    public static Exception unwrap(Throwable t) {
-        assert t != null;
-
-        while (true) {
-            if (t instanceof Error)
-                throw (Error)t;
-
-            if (t instanceof GridClosureException) {
-                t = ((GridClosureException)t).unwrap();
-
-                continue;
-            }
-
-            return (Exception)t;
-        }
-    }
-
-    /**
-     * Casts the passed {@code Throwable t} to {@link IgniteCheckedException}.<br>
-     * If {@code t} is a {@link GridClosureException}, it is unwrapped and then cast to {@link IgniteCheckedException}.
-     * If {@code t} is an {@link IgniteCheckedException}, it is returned.
-     * If {@code t} is not a {@link IgniteCheckedException}, a new {@link IgniteCheckedException} caused by {@code t}
-     * is returned.
-     *
-     * @param t Throwable to cast.
-     * @return {@code t} cast to {@link IgniteCheckedException}.
-     */
-    public static IgniteCheckedException cast(Throwable t) {
-        assert t != null;
-
-        t = unwrap(t);
-
-        return t instanceof IgniteCheckedException
-            ? (IgniteCheckedException)t
-            : new IgniteCheckedException(t);
     }
 
     /**
